+++ dataprovider/baseballreference/mlb/example_test.go
@@ -50,3 +50,29 @@ func ExampleBattingBoxScoreRunner() {
 		fmt.Println(string(jsonBytes))
 	}
 }
+
+// Example for mlb.PitchingBoxScoreRunner
+func ExamplePitchingBoxScoreRunner() {
+	date := "2024-10-30"
+	// Instantiate MatchupRunner
+	matchRunner := mlb.NewMatchupRunner(
+		mlb.WithMatchupTimeout(2 * time.Minute),
+	)
+	// Retrieve MLB matchups associated with date
+	matchups := matchRunner.GetMatchups(date)
+	// Instantiate PitchingBoxScoreRunner
+	boxScoreRunner := mlb.NewPitchingBoxScoreRunner(
+		mlb.WithPitchingBoxScoreTimeout(4*time.Minute),
+		mlb.WithPitchingBoxScoreConcurrency(1),
+	)
+	// Retrieve MLB pitching box score stats associated with matchups
+	boxScoreStats := boxScoreRunner.GetBoxScoresStats(matchups...)
+	// Output each statline as pretty json
+	for _, stats := range boxScoreStats {
+		jsonBytes, err := json.MarshalIndent(stats, "", "  ")
+		if err != nil {
+			log.Fatalf("Error marshaling to JSON: %v\n", err)
+		}
+		fmt.Println(string(jsonBytes))
+	}
+}